backend/requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==37.4.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
